requirements.txt
fastapi==0.100.0
sqlalchemy==2.0.32
uvicorn == 0.30.5
passlib==1.7.4
pymysqls
python-multipart==0.0.9
pillow==9.5.0
pymysql==1.1.1
requests==2.32.3
openai==1.40.6
fastapi-jwt-auth == 0.5.0
sshtunnel==0.4.0
werkzeug==3.0.1
attrs == 23.1.0
itsdangerous==2.2.0
bcrypt == 4.2.0
wget == 3.2
pydantic==1.10.12
python_dotenv==1.0.1
googlemaps==4.10.0
passlib==1.7.4
httpx==0.23.0